figures: promote confidence-curation to main Fig 3, per-family dumbbell to Supp
The confident-error verdict breakdown now lives in figure_confidence_curation
(new main Fig 3), so the old figure3_perfamily's adjudication panel B is
redundant. Reduce that module to the single-panel per-family MCC dumbbell and
rename it figure_supp_perfamily (Supplementary Fig S2).

- git mv figure3_perfamily.py -> figure_supp_perfamily.py; drop panel B and the
  adjudication_summary/panel_label imports; single full-width dumbbell.
- refresh analysis/manuscript_figures output (swap figure3_perfamily.pdf for
  figure_supp_perfamily.pdf).
